requirements.txt
ubatch
aiohttp
pandas
psycopg2
matplotlib
numpy
threaded
python-dotenv
alembic==1.4.3
click==7.1.2
fastapi==0.63.0
flake8==3.8.4
h11==0.11.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pydantic==1.7.3
pydocstyle==5.1.1
pyflakes==2.2.0
pylama==7.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing==3.7.4.3
uvicorn==0.13.2
